+++ middlewares/validationMiddlewares.js
@@ -1,0 +1,22 @@
+const Joi = require("joi");
+
+module.exports = {
+  addContactValidation: (req, res, next) => {
+    const schema = Joi.object({
+      name: Joi.string().min(3).max(30).required(),
+      email: Joi.string()
+        .email({
+          minDomainSegments: 2,
+        })
+        .required(),
+      phone: Joi.string().required(),
+    });
+
+    const validationResult = schema.validate(req.body);
+    if (validationResult.error) {
+      return res.status(400).json({ message: validationResult.error.details });
+    }
+
+    next();
+  },
+};
