+++ models/contacts.js
@@ -1,7 +1,6 @@
 const fs = require("fs/promises");
 const path = require("path");
 const { uid } = require("uid");
-// const Joi = require("joi");
 
 const contactsPath = path.join(__dirname, "contacts.json");
