+++ models/contacts.json
@@ -55,26 +55,8 @@
   },
   {
     "id": "10",
-    "name": "Alec Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
-  },
-  {
-    "id": "d96cd3da617",
     "name": "Nick Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
-  },
-  {
-    "id": "c8aa2c87442",
-    "name": "Nickrrr Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "688"
-  },
-  {
-    "id": "83e3c65e07e",
-    "name": "Nickrrr Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(715)598-5792"
+    "phone": "1234567899"
   }
 ]
